# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Parallel AI Coding Orchestrator API",
    "description": "\n## Week 2 MVP - Manager AI Integration API\n\nEnterprise-grade REST API for orchestrating and monitoring parallel AI coding workflows.\n\n### Key Features\n- **Supervisor API**: Monitor and control Claude Code worker instances\n- **Resource Management**: Hierarchical resource allocation with depth-based quotas\n- **Job Orchestration**: Manage job lifecycle with state machine validation\n- **Authentication**: JWT-based authentication with scope-based authorization\n\n### Architecture\n- **Database**: SQLite (development) / PostgreSQL (production)\n- **Authentication**: JWT Bearer tokens with scopes\n- **State Management**: Validated state machines for workers and jobs\n- **Testing**: 44 integration tests with 29% code coverage\n\n### Quick Start\n1. Generate test token: Use `/api/auth/token` endpoint\n2. Access protected endpoints with `Authorization: Bearer <token>` header\n3. Explore interactive docs at `/docs` or `/redoc`\n\n### Support\n- **GitHub**: [parallel-coding](https://github.com/your-org/parallel-coding)\n- **Documentation**: [Wiki](https://github.com/your-org/parallel-coding/wiki)\n",
    "contact": {
      "name": "Parallel AI Development Team",
      "email": "dev@parallel-ai.example.com"
    },
    "license": {
      "name": "MIT"
    },
    "version": "2.0.0-week2-mvp"
  },
  "paths": {
...
    "/api/v1/supervisor/": {
      "get": {
        "tags": [
          "supervisor"
        ],
        "summary": "List Supervisors",
        "description": "List active supervisors with optional filtering and pagination.",
        "operationId": "list_supervisors_api_v1_supervisor__get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "status_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "spawning",
                    "running",
                    "terminated",
                    "error"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Filter"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListSupervisorsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/supervisor/workers": {
      "get": {
        "tags": [
          "supervisor"
        ],
        "summary": "List all workers",
        "description": "Retrieve a paginated list of workers with optional filtering.\n\n    **Filters:**\n    - `workspace_id`: Filter by workspace\n    - `status`: Filter by worker status (IDLE, RUNNING, PAUSED, COMPLETED, FAILED, TERMINATED)\n\n    **Pagination:**\n    - `limit`: Maximum number of results (1-100, default: 50)\n    - `offset`: Number of results to skip (default: 0)\n\n    **Required Scope:** `supervisor:read`\n\n    **Example Response:**\n    ```json\n    {\n      \"workers\": [\n        {\n          \"id\": \"worker-123\",\n          \"workspace_id\": \"ws-001\",\n          \"status\": \"RUNNING\",\n          \"created_at\": \"2025-10-28T10:00:00Z\",\n          \"updated_at\": \"2025-10-28T10:05:00Z\"\n        }\n      ],\n      \"total\": 1,\n      \"limit\": 50,\n      \"offset\": 0\n    }\n    ```",
        "operationId": "list_workers_api_supervisor_workers_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by workspace ID",
              "title": "Workspace Id"
            },
            "description": "Filter by workspace ID"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/WorkerStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by worker status",
              "title": "Status"
            },
            "description": "Filter by worker status"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Maximum number of results",
              "default": 50,
              "title": "Limit"
            },
            "description": "Maximum number of results"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of results to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of results to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "List of workers matching the filter criteria",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkerListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/supervisor/workers/{worker_id}": {
      "get": {
        "tags": [
          "supervisor"
        ],
        "summary": "Get Worker",
        "description": "Get a single worker by id.",
        "operationId": "get_worker_api_supervisor_workers__worker_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "worker_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Worker Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/supervisor/workers/{worker_id}/pause": {
      "post": {
        "tags": [
          "supervisor"
        ],
        "summary": "Pause Worker",
        "description": "Pause worker execution.",
        "operationId": "pause_worker_api_supervisor_workers__worker_id__pause_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "worker_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Worker Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/supervisor/workers/{worker_id}/resume": {
      "post": {
        "tags": [
          "supervisor"
        ],
        "summary": "Resume Worker",
        "description": "Resume worker execution.",
        "operationId": "resume_worker_api_supervisor_workers__worker_id__resume_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "worker_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Worker Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/supervisor/workers/{worker_id}/terminate": {
      "post": {
        "tags": [
          "supervisor"
        ],
        "summary": "Terminate Worker",
        "description": "Terminate a worker.",
        "operationId": "terminate_worker_api_supervisor_workers__worker_id__terminate_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "worker_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Worker Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/workers/{worker_id}": {
      "get": {
        "summary": "Get Worker Info",
        "description": "Get information about a specific worker.\n\nArgs:\n    worker_id: Worker identifier (e.g., \"worker_001\")\n\nReturns:\n    Worker details including dialogue status\n\nRaises:\n    HTTPException: If worker not found",
        "operationId": "get_worker_info_api_v1_workers__worker_id__get",
        "parameters": [
          {
            "name": "worker_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Worker Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ListSupervisorsResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/SupervisorStatusResponse"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "total",
          "items"
        ],
        "title": "ListSupervisorsResponse"
      },
...
      "SupervisorStatusResponse": {
        "properties": {
          "supervisor_id": {
            "type": "string",
            "title": "Supervisor Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "spawning",
              "running",
              "terminated",
              "error"
            ],
            "title": "Status"
          },
          "alive": {
            "type": "boolean",
            "title": "Alive"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "uptime_secs": {
            "type": "number",
            "title": "Uptime Secs",
            "default": 0.0
          },
          "output_lines": {
            "type": "integer",
            "title": "Output Lines",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "supervisor_id",
          "status",
          "alive"
        ],
        "title": "SupervisorStatusResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkerListResponse": {
        "properties": {
          "workers": {
            "items": {
              "$ref": "#/components/schemas/WorkerResponse"
            },
            "type": "array",
            "title": "Workers"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "workers",
          "total",
          "limit",
          "offset"
        ],
        "title": "WorkerListResponse"
      },
      "WorkerResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "workspace_id",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "WorkerResponse",
        "description": "Serializable worker representation for responses."
      },
      "WorkerStatus": {
        "type": "string",
        "enum": [
          "IDLE",
          "RUNNING",
          "PAUSED",
          "COMPLETED",
          "FAILED",
          "TERMINATED"
        ],
        "title": "WorkerStatus",
        "description": "Worker lifecycle status enumeration.\n\nStates:\n    IDLE: Worker created but not yet started\n    RUNNING: Worker actively executing task\n    PAUSED: Worker temporarily suspended\n    COMPLETED: Worker finished successfully\n    FAILED: Worker encountered fatal error\n    TERMINATED: Worker manually terminated"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}